Add new build target 'minikernel':
./rootfs minikernel

will download external helper script,
which compiles a kernel and starts qemu
(if possible in kvm mode) together with
the already generated initital ramdisk.

all tempfiles are in already provided tmpfs
and lost at the very end during unmount and exit.

also add a logfile for all targets, which
can be used to better spot errors.