fixes bash build. fixes: https://github.com/fosslinux/live-bootstrap/issues/99
in my usecase (qemu, kernel 3.18.140) around 50% of
all builds are failing with a wrong binary hash.
the culprit is a varying generation of 'pipesize.h' during:

```
/bin/sh ./psize.sh > pipesize.h
```

normally generated 'pipesize.h' looks like:
```
/*
* pipesize.h
*
* This file is automatically generated by psize.sh
* Do not edit!
*/

 #define PIPESIZE 65536
```

in the error case the last line has 512 instead of 65536.
With this change we use a fixed output of the large PIPESIZE.